UI: Rename Utilities.cpp -> Utils.cpp
Matches naming convention of other "utility" files.
