+++ Ladybird/Android/src/main/cpp/RequestServerService.cpp
@@ -7,7 +7,7 @@
 
 #include <AK/LexicalPath.h>
 #include <AK/OwnPtr.h>
-#include <Ladybird/Utilities.h>
+#include <Ladybird/Utils.h>
 #include <LibCore/ArgsParser.h>
 #include <LibCore/EventLoop.h>
 #include <LibCore/LocalServer.h>
