+++ Ladybird/Android/src/main/cpp/WebContentService.cpp
@@ -10,7 +10,7 @@
 #include <Ladybird/FontPlugin.h>
 #include <Ladybird/HelperProcess.h>
 #include <Ladybird/ImageCodecPlugin.h>
-#include <Ladybird/Utilities.h>
+#include <Ladybird/Utils.h>
 #include <LibAudio/Loader.h>
 #include <LibCore/ArgsParser.h>
 #include <LibCore/EventLoop.h>
