+++ Ladybird/WebDriver/main.cpp
@@ -5,7 +5,7 @@
  */
 
 #include <AK/Platform.h>
-#include <Ladybird/Utilities.h>
+#include <Ladybird/Utils.h>
 #include <LibCore/ArgsParser.h>
 #include <LibCore/Directory.h>
 #include <LibCore/EventLoop.h>
