+++ Meta/gn/secondary/Ladybird/WebDriver/BUILD.gn
@@ -25,7 +25,7 @@ executable("WebDriver") {
     "//Userland/Libraries/LibWebView:WebDriverServerEndpoint",
   ]
   sources = [
-    "//Ladybird/Utilities.cpp",
+    "//Ladybird/Utils.cpp",
     "//Userland/Services/WebDriver/Client.cpp",
     "//Userland/Services/WebDriver/Session.cpp",
     "//Userland/Services/WebDriver/WebContentConnection.cpp",
